src/main/requirements.txt
Django==3.2.14
#psycopg2-binary==2.8.6
psycopg2-binary==2.9.3
django-import-export==2.5.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
requests==2.28.1
django-environ==0.4.5
pandas==1.1.5
numpy==1.19.5
python-dateutil==2.8.2
six==1.16.0
#django-keycloak-auth==0.9.3
django-cors-headers==3.9.0
